Ajuste de tarefas e downloads

+++ Gert.Site/Controllers/AlunoController.cs
@@ -22,10 +22,14 @@ namespace Gert.Site.Controllers
         {
             try
             {
+                var semestre = (DateTime.Now.Month <= 6 ? 1 : 2);
                 var usuario = LoginUtils.Usuario;
-                var disciplinasAluno = GertDbFactory.Instance.DisciplinaAlunoRepository.FindByIdAluno(usuario.Pessoa.Id);
+                
                 var disciplinas = GertDbFactory.Instance.DisciplinaRepository.FindAll();
-                disciplinas = disciplinas.Where(w => w.Ativa).ToList();
+                disciplinas = disciplinas.Where(w => w.Ativa && w.Semestre == semestre && w.Ano == DateTime.Now.Year).ToList();
+
+                var disciplinasAluno = GertDbFactory.Instance.DisciplinaAlunoRepository.FindByIdAluno(usuario.Pessoa.Id);
+                disciplinasAluno = disciplinasAluno.Where(w => w.Disciplina.Ativa && w.Disciplina.Semestre == semestre && w.Disciplina.Ano == DateTime.Now.Year).ToList();
 
                 var naoMatriculadas = new List<Disciplina>();
                 var matriculadas = new List<Disciplina>();
@@ -54,15 +58,22 @@ namespace Gert.Site.Controllers
         {
             try
             {
-                if(tipo == 0)
+                var usuario = LoginUtils.Usuario;
+                var disciplinasAluno = GertDbFactory.Instance.DisciplinaAlunoRepository.FindByIdAluno(usuario.Pessoa.Id);
+
+                if (tipo == 0)
                 {
-                    var usuario = LoginUtils.Usuario;
-                    var disciplinasAluno = GertDbFactory.Instance.DisciplinaAlunoRepository.FindByIdAluno(usuario.Pessoa.Id);
+                    
 
                     var matriculada = disciplinasAluno.FirstOrDefault(f => f.Disciplina.Id == id);
                     if (matriculada == null)
                     {
                         return Json(new { success = true, Message = "Ok" });
+                    }
+
+                    if(matriculada.Disciplina.Tarefas.Count > 0)
+                    {
+                        return Json(new { success = false, error = true, Message = "Não é possível remover uma disciplina que já contenha tarefas cadastradas." });
                     }
 
                     GertDbFactory.Instance.DisciplinaAlunoRepository.Delete(matriculada);
@@ -72,10 +83,14 @@ namespace Gert.Site.Controllers
 
                     if (disciplina == null)
                     {
-                        return Json(new { success = true, Message = "Ok" });
+                        return Json(new { success = false, Message = "Disciplina não encontrada." });
                     }
 
-                    var usuario = LoginUtils.Usuario;
+                    var matriculada = disciplinasAluno.FirstOrDefault(f => f.Disciplina.Id == id);
+                    if (matriculada != null)
+                    {
+                        return Json(new { success = true, Message = "Ok" });
+                    }
 
                     var matricula = new DisciplinaAluno()
                     {
@@ -93,6 +108,43 @@ namespace Gert.Site.Controllers
             catch (Exception ex)
             {
                 return Json(new { success = false, Message = "Não foi possível gravar a discplina!" + ex.Message });
+            }
+        }
+
+        [HttpPost]
+        public PartialViewResult Perfil()
+        {
+            try
+            {
+                var user = LoginUtils.Usuario;
+                user.Pessoa.Usuario = user;
+
+                return PartialView("_Perfil", user.Pessoa);
+            }catch(Exception ex)
+            {
+                return PartialView("Error", new HandleErrorInfo(ex, "Aluno", "Perfil"));
+            }
+        }
+
+        [HttpPost]
+        public ActionResult AlterarAluno(Pessoa pessoa, Usuario usuario)
+        {
+            try
+            {
+                var user = LoginUtils.Usuario;
+                user.Login = usuario.Login;
+                user.Senha = usuario.Senha;
+
+                user.Pessoa = pessoa;
+                GertDbFactory.Instance.UsuarioRepository.Save(user);
+
+                LoginUtils.Logar(user.Login, user.Senha);
+
+                return RedirectToAction("Index");
+            }
+            catch (Exception ex)
+            {
+                return PartialView("Error", new HandleErrorInfo(ex, "Aluno", "AlterarAluno"));
             }
         }
     }
